Feat: Introduce `ByteVecGen` for structs generates bytecode
