+++ src/asm/generate/mod.rs
@@ -1,4 +1,8 @@
+use crate::asm::generate::byte_vec::ByteVecImpl;
+use crate::asm::generate::symbol::SymbolTable;
+
 pub mod annotation;
+pub(crate) mod attribute;
 pub mod byte_vec;
 pub mod class;
 pub mod constant_value;
@@ -12,3 +16,7 @@ pub mod record;
 pub mod signature;
 pub(crate) mod symbol;
 pub mod types;
+
+pub(crate) trait ByteVecGen {
+    fn put(&self, byte_vec: &mut ByteVecImpl, symbol_table: &mut SymbolTable);
+}
